+++ pytket/binders/circuit/main.cpp
@@ -323,6 +323,12 @@ PYBIND11_MODULE(circuit, m) {
           ":math:`e^{-\\frac{i\\pi\\alpha}{2} P}` of a tensor "
           ":math:`P` of Pauli operations.")
       .value(
+          "TermSequenceBox", OpType::TermSequenceBox,
+          "An unordered collection of Pauli exponentials "
+          "that can be synthesised in any order, causing a "
+          "change in the unitary operation. Synthesis order "
+          "depends on the synthesis strategy chosen only.")
+      .value(
           "QControlBox", OpType::QControlBox,
           "An arbitrary n-controlled operation")
       .value(