+++ pytket/docs/changelog.rst
@@ -17,6 +17,9 @@ Features:
   ``BackendResult.get_probability_distribution()`` and to the constructor of a
   ``ProbabilityDistribution``, defaulting to zero. (Previously probabilities
   below 1e-10 were by default treated as zero.)
+* Add python binding for ``UnitaryRevTableau``.
+* Add ``TermSequenceBox``, for circuit synthesis of a series of Pauli 
+  Exponentials, where the ordering of terms can be changed.
 * Add ``circuit_name`` property to ``CircBox``.
 
 Fixes: